feat: Se realiza el cambio a consumo de resoluciones y la aplicacion de la libreria UTILIDADES, de manera que cambian varias cosas del modelo y asi mismo se agrega la variable de entorno hacia resoluciones.

+++ models/pago_mensual.go
@@ -6,7 +6,7 @@ type PagoMensual struct {
 	VigenciaContrato  float64            
 	Mes               float64            
 	Persona           string             
-	EstadoPagoMensualId float64  
+	EstadoPagoMensualId int  
 	Responsable       string  
 	FechaCreacion	  string
 	FechaModificacion string
